Come on in! The water's warm!
