+++ skxray/api/fluorescence.py
@@ -41,17 +41,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 # import fitting models
-from ..fitting.api import (
-    ConstantModel, LinearModel, QuadraticModel, ParabolicModel,
-    PolynomialModel, VoigtModel, PseudoVoigtModel, Pearson7Model,
-    StudentsTModel, BreitWignerModel, GaussianModel, LorentzianModel,
-    LognormalModel, DampedOscillatorModel, ExponentialGaussianModel,
-    SkewedGaussianModel, DonaichModel, PowerLawModel, ExponentialModel,
-    StepModel, RectangleModel, Lorentzian2Model, ComptonModel, ElasticModel
-)
 
 # import Element objects
-from ..constants.api import XrfElement, emission_line_search
 
 # import background subtraction
-from ..fitting.background import snip_method
