+++ skxray/core/constants/__init__.py
