+++ skxray/core/constants/api.py
