+++ skxray/core/constants/basic.py
