+++ skxray/core/image.py
