+++ skxray/core/roi.py
