+++ skxray/tests/core/test_constants/test_xrs.py
@@ -44,7 +44,7 @@ import numpy as np
 from numpy.testing import (assert_array_equal, assert_array_almost_equal)
 from nose.tools import assert_equal
 
-from skxray.constants.xrs import (HKL,
+from skxray.core.constants.xrs import (HKL,
                                   calibration_standards)
 from skxray.core.utils import q_to_d, d_to_q
