+++ skxray/tests/core/test_correlation.py
@@ -2,9 +2,6 @@
 # Copyright (c) 2014, Brookhaven Science Associates, Brookhaven        #
 # National Laboratory. All rights reserved.                            #
 #                                                                      #
-# @author: Li Li (lili@bnl.gov)                                        #
-# created on 08/19/2014                                                #
-#                                                                      #
 # Redistribution and use in source and binary forms, with or without   #
 # modification, are permitted provided that the following conditions   #
 # are met:                                                             #
@@ -35,11 +32,59 @@
 # IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE   #
 # POSSIBILITY OF SUCH DAMAGE.                                          #
 ########################################################################
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+import logging
+
+import numpy as np
+from numpy.testing import (assert_array_almost_equal,
+                           assert_almost_equal)
+from skimage import data
+
+import skxray.core.correlation as corr
+import skxray.core.roi as roi
+from skxray.testing.decorators import skip_if
+
+logger = logging.getLogger(__name__)
 
 
-# import * is only allowed at the module level
-from skxray.constants.api import *
+# It is unclear why this test is so slow. Can we speed this up at all?
+@skip_if(True)
+def test_correlation():
+    num_levels = 4
+    num_bufs = 8  # must be even
+    num_qs = 2  # number of interested roi's (rings)
+    img_dim = (50, 50)  # detector size
 
-if __name__ == '__main__':
-    import nose
-    nose.runmodule(argv=['-s', '--with-doctest'], exit=False)
+    roi_data = np.array(([10, 20, 12, 14], [40, 10, 9, 10]),
+                        dtype=np.int64)
+
+    indices = roi.rectangles(roi_data, img_dim)
+
+    img_stack = np.random.randint(1, 5, size=(500, ) + img_dim)
+
+    g2, lag_steps = corr.multi_tau_auto_corr(num_levels, num_bufs, indices,
+                                             img_stack)
+
+    assert_array_almost_equal(lag_steps,  np.array([0, 1, 2, 3, 4, 5, 6, 7, 8,
+                                                   10, 12, 14, 16, 20, 24, 28,
+                                                   32, 40, 48, 56]))
+
+    assert_array_almost_equal(g2[1:, 0], 1.00, decimal=2)
+    assert_array_almost_equal(g2[1:, 1], 1.00, decimal=2)
+
+    coins = data.camera()
+    coins_stack = []
+
+    for i in range(500):
+        coins_stack.append(coins)
+
+    coins_mesh = np.zeros_like(coins)
+    coins_mesh[coins < 30] = 1
+    coins_mesh[coins > 50] = 2
+
+    g2, lag_steps = corr.multi_tau_auto_corr(num_levels, num_bufs, coins_mesh,
+                                             coins_stack)
+
+    assert_almost_equal(True, np.all(g2[:, 0], axis=0))
+    assert_almost_equal(True, np.all(g2[:, 1], axis=0))
